1. Добавлена проверка на невалидные входные данные (не цифры) 2. Исправлена проверка с дробными числами

+++ 1 лаба/Triangle/Triangle/Triangle.cpp
@@ -2,7 +2,7 @@
 #include "Triangle.h"
 
 
-CTriangle::CTriangle(unsigned a, unsigned b, unsigned c)
+CTriangle::CTriangle(float a, float b, float c)
 	:a(a)
 	,b(b)
 	,c(c)
